requirements.txt
yfinance
pandas
numpy
scikit-learn
xgboost
textblob
newsapi-python
streamlit
plotly
